refactoring model wrapper
