+++ oslo/torch/nn/parallel/tensor_parallel/__init__.py
@@ -2,5 +2,6 @@ from oslo.torch.nn.parallel.tensor_parallel.mapping import Column, Row, Update, 
 from oslo.torch.nn.parallel.tensor_parallel.tensor_parallel import (
     TensorParallel
 )
+from oslo.torch.nn.parallel.tensor_parallel._base_wrapper import BaseTensorParallelWrapper
 
-__ALL__ = [TensorParallel, Column, Row, Update, Head]
+__ALL__ = [TensorParallel, Column, Row, Update, Head, BaseTensorParallelWrapper]
